naming refactoring for event queue
